Finish structure and style of AggregateMacSettingsIndicators
