+++ frontend/pages/ManageControlsPage/MacOSSettings/MacOSSettings.tsx
@@ -2,10 +2,9 @@ import React from "react";
 import { Params } from "react-router/lib/Router";
 
 import SideNav from "pages/admin/components/SideNav";
-import { IAggregateMacSettingsStatus } from "interfaces/mdm";
 
 import MAC_OS_SETTINGS_NAV_ITEMS from "./MacOSSettingsNavItems";
-import AggregateMacSettings from "./AggregateMacSettings";
+import AggregateMacSettingsIndicators from "./AggregateMacSettingsIndicators";
 
 const baseClass = "mac-os-settings";
 
@@ -23,21 +22,12 @@ const MacOSSettings = ({ params }: IMacOSSettingsProps) => {
 
   const CurrentCard = currentFormSection.Card;
 
-  const dummyData: IAggregateMacSettingsStatus = {
-    latest: 100,
-    pending: 100,
-    failing: 100,
-  };
-
   return (
     <div className={baseClass}>
       <p className={`${baseClass}__description`}>
         Remotely enforce settings on macOS hosts assigned to this team.
       </p>
-      <AggregateMacSettings
-        // team_id={params.team_id}
-        aggregateProfileData={dummyData}
-      />
+      <AggregateMacSettingsIndicators />
       <SideNav
         className={`${baseClass}__side-nav`}
         navItems={MAC_OS_SETTINGS_NAV_ITEMS}
