+++ frontend/pages/ManageControlsPage/MacOSSettings/_styles.scss
@@ -1,14 +1,51 @@
 .mac-os-settings {
   font-size: $x-small;
+  margin-top: $pad-xxlarge;
+  display: flex;
+  flex-direction: column;
+  gap: $pad-xxlarge;
 
   &__description {
-    margin: $pad-xxlarge 0;
+    margin: 0;
   }
 
   &__side-nav {
-
     .side-nav__nav-list {
-      top: 0
+      top: 0;
     }
+  }
+}
+
+.aggregate-mac-settings-indicators {
+  display: flex;
+  height: 94px;
+  border-top: 1px solid #e2e4ea;
+  border-bottom: 1px solid #e2e4ea;
+  border-left: 1px solid #e2e4ea;
+  border-radius: 6px;
+
+  .aggregate-mac-settings-indicator {
+    flex-grow: 1;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    border-right: 1px solid #e2e4ea;
+
+    // but not the tooltip text
+    .tooltip {
+      font-weight: $bold;
+      flex-direction: column;
+    }
+
+    a {
+      font-weight: $regular;
+    }
+  }
+
+  .aggregate-mac-settings-indicator:last-child {
+    border-top-right-radius: 6px;
+    border-bottom-right-radius: 6px;
   }
 }
